// dtApi/src/app/modules/admin/test/test.interfaces.ts
export interface DialogData {
    data: any;
    description: any;
}
export interface Subject {
    subject_id: number;
    subject_name: string;
    subject_description: string;
}
export interface Test {
    test_id: number;
    test_name: string;
    subject_id: number;
    tasks: number;
    time_for_test: number;
    enabled: boolean;
    attempts: number;
}
export interface ServiceResponse {
    tests: [];
    subjects: [];
}